WIP: split DetectorsApiService/DatasetsApiService into per-domain slices - pre-test checkpoint
